pyproject.toml
[project]
name = "gen-tree"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0",
    "django-livereload-server>=0.5.1",
    "django-neomodel>=0.2.0",
]
